+++ cpp_project/coords.h
@@ -7,20 +7,20 @@
 class Coords
 {
 public:
-    Coords(unsigned int x, unsigned int y) : xCoord(x), yCoord(y) {}
+    Coords(unsigned short int x, unsigned short int y) : xCoord(x), yCoord(y) {}
 
-    unsigned int x() { return xCoord; } // not getX() because this method will be used a lot and x() seems clearer.
-    unsigned int y() { return yCoord; } // not getY() because this method will be used a lot and y() seems clearer.
+    unsigned short int x() { return xCoord; } // not getX() because this method will be used a lot and x() seems clearer.
+    unsigned short int y() { return yCoord; } // not getY() because this method will be used a lot and y() seems clearer.
 
     void setX(unsigned int x) { xCoord = x; }
     void setY(unsigned int y) { yCoord = y; }
 
-    void addX(int x) { xCoord += x; }
-    void addY(int y) { yCoord += y; }
+    void addX(short int x) { xCoord += x; }
+    void addY(short int y) { yCoord += y; }
 
 private:
-    unsigned int xCoord;
-    unsigned int yCoord;
+    unsigned short int xCoord;
+    unsigned short int yCoord;
 };
 
 #endif // COORDS_H
