+++ cpp_project/vampire.cpp
@@ -3,7 +3,7 @@
 #include <math.h>
 #include <numbers>
 
-Vampire::Vampire(Coords *hitBox, Coords *position, unsigned int speed, unsigned int maxHealth, unsigned int currentHealth, unsigned int damage, Coords* target):
+Vampire::Vampire(Coords *hitBox, Coords *position, unsigned short int speed, unsigned short int maxHealth, unsigned short int currentHealth, unsigned short int damage, Coords* target):
     Enemy(hitBox, position, speed, maxHealth, currentHealth, target), damage(damage)
 {
 }
@@ -39,7 +39,7 @@ void Vampire::moveOnlyY()
     position->addY(std::round(target->y() > position->y() ? moveAllowance : -1 * moveAllowance));
 }
 
-void Vampire::modifyHealth(int amount)
+void Vampire::modifyHealth(short int amount)
 {
     currentHealth += amount;
 }
